requirements.txt
python-telegram-bot==21.0
mcstatus==11.1.1
psutil==5.9.8
